requirements.txt
numpy
pyarrow
faiss-cpu